FIX: reliable method for changing origin host
Fix an error which occurs most notably when you
attempt to `gram assign` one user, and then try to
`gram assign` again but to another user: gram
previously did a find-replace for ".com", so once
gram already changed the origin host to
"github-username" there's no longer a .com and no
longer the ability to change the origin based on
the old method. The new method implemented (the
change) simply changes the git config origin host
value between @ and : which we know can always
blindly be updated to `github-someuser`.